+++ HISTORY.rst
@@ -8,6 +8,12 @@ Changelog
 - disallow logins from non-backend URLs, if set in Security panel; tweaks to field descriptions
   [CorySanin]
 
+- improve export and import scripts
+  [tkimnguyen]
+
+- tweak find-broken-links.py script
+  [tkimnguyen]
+
 - Changed collective.documentviewer dep. version
   [obct537]